requirements.txt
streamlit
PyMuPDF
python-docx
sentence-transformers
spacy
rapidfuzz

